search with product name ordered by price

+++ ecommerce/main/views.py
@@ -1,7 +1,7 @@
 from .forms import RegisterForm
 from django.contrib.auth.decorators import login_required
 from django.contrib.auth import login, logout, authenticate
-from .models import Product , Cart , Order , OrderItem
+from .models import Product, Cart, Order, OrderItem
 from django.views.decorators.csrf import csrf_exempt
 from django.http import JsonResponse
 from django.core import serializers
@@ -52,19 +52,20 @@ def get_products(request):
         products = Product.objects.all()
         data = serializers.serialize('json', products)
         return JsonResponse({
-            "products":json.loads(data)
+            "products": json.loads(data)
         })
     else:
         return JsonResponse({
-        "message": "Request Must be get"
+            "message": "Request Must be get"
         })
     
+
 def get_products_ordered(request):
     if request.method == "GET":
         products = Product.objects.all().order_by('price')
         data = serializers.serialize('json', products)
         return JsonResponse({
-            "products":json.loads(data)
+            "products": json.loads(data)
         })
     else:
         return JsonResponse({
@@ -93,15 +94,15 @@ def log_in(request):
         if user is not None:
             login(request, user)
             return JsonResponse({
-        "message": "User logged in Succesfully"
+                  "message": "User logged in Succesfully"
         })
         else:
-         return JsonResponse({
-        "message": "Error occured"
+            return JsonResponse({
+                "message": "Error occured"
         })
     else:
         return JsonResponse({
-        "message": "Request must be Post"
+              "message": "Request must be Post"
         })
     
 @login_required(login_url="/")
@@ -112,19 +113,9 @@ def add_to_cart(request):
         quantity = request.POST['quantity']
         product = Product.objects.get(pk=productId)
         if (product):
-            #serializedProduct = serializers.serialize('json', product)
-            #jsonProduct = json.loads(serializedProduct)
-            #productPrice = jsonProduct[0].get("fields").get("price")
-            #productName = jsonProduct[0].get("fields").get("name")
-            #productCreatedAt = jsonProduct[0].get("fields").get("created_at")
-            #productUpdatedAt = jsonProduct[0].get("fields").get("updated_at")
-            #productPK = jsonProduct[0].get("fields").get("pk")
-            #productInstance = Product(pk=productPK, name=productName, price=productPrice, 
-            #                          created_at=productCreatedAt,
-            #                          updated_at=productUpdatedAt)
             Cart.objects.create(customer=request.user, product=product, quantity = int(quantity) ,price = int(quantity) * product.price)
             return JsonResponse({
-        "message": "Added to cart successfully"
+                "message": "Added to cart successfully"
         })
 
         else:
@@ -134,22 +125,22 @@ def add_to_cart(request):
         
     else:
         return JsonResponse({
-        "message": "Request must be Post"
+                "message": "Request must be Post"
         })
     
-@login_required(login_url="/")
+@login_required(login_url="/login")
 @csrf_exempt
 def get_cart(request):
     if request.method == "GET":
         cart = Cart.objects.filter(customer=request.user.id)
         data = serializers.serialize('json', cart)
         return JsonResponse({
-            "products":json.loads(data)
+            "products": json.loads(data)
         })
         
     else:
         return JsonResponse({
-        "message": "Request must be gET"
+                "message": "Request must be gET"
         })
 
 @login_required(login_url="/")
@@ -159,7 +150,7 @@ def get_orders(request):
         orders = Order.objects.filter(customer=request.user.id)
         data = serializers.serialize('json', orders)
         return JsonResponse({
-            "products":json.loads(data)
+            "products": json.loads(data)
         })
     else:
         return JsonResponse({
@@ -171,8 +162,7 @@ def get_orders(request):
 def get_product_by_name(request):
     if request.method == "GET":
         productName = request.GET.get('q','')
-        #orders = Product.objects.filter(name__contains=productName)
-        orders = Product.objects.filter(Q(name__contains=productName.lower()) | Q(name__contains=productName.upper()))
+        orders = Product.objects.filter(Q(name__contains=productName.lower()) | Q(name__contains=productName.upper())).order_by('price')
         data = serializers.serialize('json', orders)
         return JsonResponse({
             "products":json.loads(data)
@@ -189,7 +179,7 @@ def get_orders_details(request):
         orders = OrderItem.objects.filter(customer=request.user.id)
         data = serializers.serialize('json', orders)
         return JsonResponse({
-            "orders":json.loads(data)
+            "orders": json.loads(data)
         })
     else:
         return JsonResponse({
@@ -201,26 +191,36 @@ def get_orders_details(request):
 def create_order(request):
     if request.method == "POST":
         cartItems = Cart.objects.filter(customer=request.user.id)
-        cartItemsSerialized = serializers.serialize('json', cartItems)
-        cartItemsJson = json.loads(cartItemsSerialized)
-        itemsSet = set()
-        totalPrice = 0
-        for item in cartItemsJson:
-            product = Product.objects.get(pk=int(item.get("fields").get("product")))
-            itemsSet.add(product.name)
-            totalPrice += item.get("fields").get("price")
-        order = Order(customer=request.user, products=list(itemsSet) , totalPrice=totalPrice)
-        order.save()
-        for item in cartItemsJson:
-            product = Product.objects.get(pk=item.get("fields").get("product"))
-            order_item = OrderItem(customer=request.user, product=product, order=order,
-                                   quantity=item.get("fields").get("quantity"),price=item.get("fields").get("price") )
-            order_item.save()
-        Cart.objects.filter(customer=request.user.id).delete()
-        return JsonResponse({
-            "message": "Order Saved ,Cart is empty now"
-        })
+        if not cartItems:
+            raise IndexError("Can't order while Cart is empty")
+        else:
+            cartItemsSerialized = serializers.serialize('json', cartItems)
+            cartItemsJson = json.loads(cartItemsSerialized)
+            itemsSet = set()
+            totalPrice = 0
+            for item in cartItemsJson:
+                product = Product.objects.get(pk=int(item.get("fields")
+                                                     .get("product")))
+                itemsSet.add(product.name)
+                totalPrice += item.get("fields").get("price")
+            order = Order(customer=request.user, products=list(itemsSet) ,
+                           totalPrice=totalPrice)
+            order.save()
+            for item in cartItemsJson:
+                product = Product.objects.get(pk=item.get("fields")
+                                              .get("product"))
+                order_item = OrderItem(customer=request.user, product=product, 
+                                       order=order,
+                                       quantity=item.get("fields")
+                                       .get("quantity"), price=item
+                                       .get("fields")
+                                       .get("price"))
+                order_item.save()
+            Cart.objects.filter(customer=request.user.id).delete()
+            return JsonResponse({
+                "message": "Order Saved ,Cart is empty now"
+                                })
     else:
         return JsonResponse({
-        "message": "Request must be Post"
+            "message": "Request must be Post"
         })
